Moved test cases to other file
