+++ sampleParents.py
@@ -12,7 +12,7 @@ def getPrefixSums(lst):
         currSum += lst[i]
         prefixSums[i] = currSum
     totalSum = currSum
-    
+
     return prefixSums, totalSum
 
 def singleSample(prefixSums, totalSum):
@@ -38,27 +38,3 @@ def sampleBestParents(fitnesses):
         samples[i] = singleSample(prefixSums, totalSum)
 
     return samples
-
-
-def test_sampling():
-    fitnesses = [0.1, 0.9]
-    samples = sampleBestParents(fitnesses)
-    proportions = [np.sum(samples == i) for i in np.unique(samples)]
-    print(fitnesses)
-    print(proportions)
-
-    fitnesses = [0.23, 0.33, 0.10, 0.50, 0.45]
-    samples = sampleBestParents(fitnesses)
-    proportions = [np.sum(samples == i) for i in np.unique(samples)]
-    print(fitnesses)
-    print(proportions)
-
-    fitnesses = np.array([16, 10, 40, 4, 30])
-    fitnesses = fitnesses / np.sum(fitnesses)
-    samples = sampleBestParents(fitnesses)
-    proportions = [np.sum(samples == i) for i in np.unique(samples)]
-    print(fitnesses)
-    print(proportions)
-    pdb.set_trace()
-
-    return
